requirements.txt
absl-py==0.8.0
adal==1.2.2
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
azure-cognitiveservices-anomalydetector==0.2.0
azure-cognitiveservices-knowledge-qnamaker==0.1.0
azure-cognitiveservices-language-luis==0.2.0
azure-cognitiveservices-language-nspkg==3.0.1
azure-cognitiveservices-language-spellcheck==1.0.0
azure-cognitiveservices-language-textanalytics==0.2.0
azure-cognitiveservices-nspkg==3.0.1
azure-cognitiveservices-search-autosuggest==0.1.0
azure-cognitiveservices-search-customimagesearch==0.1.0
azure-cognitiveservices-search-customsearch==0.2.0
azure-cognitiveservices-search-entitysearch==1.0.0
azure-cognitiveservices-search-imagesearch==1.0.0
azure-cognitiveservices-search-newssearch==1.0.0
azure-cognitiveservices-search-nspkg==3.0.1
azure-cognitiveservices-search-videosearch==1.0.0
azure-cognitiveservices-search-visualsearch==0.2.0
azure-cognitiveservices-search-websearch==1.0.0
azure-cognitiveservices-vision-computervision==0.4.0
azure-cognitiveservices-vision-contentmoderator==1.0.0
azure-cognitiveservices-vision-customvision==1.0.0
azure-cognitiveservices-vision-face==0.4.0
azure-common==1.1.23
azure-nspkg==3.0.2
backcall==0.1.0
beautifulsoup4==4.6.3
boto3==1.9.33
botocore==1.12.33
bs4==0.0.1
certifi==2018.10.15
cffi==1.12.3
chardet==3.0.4
Click==7.0
cmake==3.15.3
cognitive-face==1.5.0
colorama==0.4.0
cryptography==2.7
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.2.4
django-extensions==2.2.3
django-werkzeug-debugger-runserver==0.3.1
dlib==19.8.1
docutils==0.14
entrypoints==0.2.3
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.1.1
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
idna==2.7
imutils==0.5.3
ipykernel==5.1.0
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
jieba==0.39
Jinja2==2.10.3
jmespath==0.9.3
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
lazy-object-proxy==1.4.1
lxml==4.4.0
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
msrest==0.6.9
msrestazure==0.6.1
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.3
notebook==5.7.0
numpy==1.17.2
oauthlib==3.1.0
opencv-python==4.1.1.26
pandas==0.23.4
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==5.4.1
plotly==4.0.0
prometheus-client==0.4.2
prompt-toolkit==2.0.9
protobuf==3.9.2
psutil==5.6.3
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.2.2
pyquery==1.4.0
pyrsistent==0.15.4
python-dateutil==2.7.3
pytz==2018.5
pywinpty==0.5.4
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.4.2
requests==2.19.1
requests-oauthlib==1.2.0
retrying==1.3.3
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.11.0
sklearn==0.0
soupsieve==1.9.3
SQLAlchemy==1.3.6
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.23
virtualenv==16.7.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.4.2
wordcloud==1.5.0
wrapt==1.11.2
wxPython==4.0.6
xgboost==0.80
